backend/requirements.txt
fastapi
uvicorn
python-dotenv

pydantic
sqlalchemy
psycopg2-binary
pgvector

openai
anthropic

pdfplumber
python-docx

tiktoken
requests

pypdf
langchain
sentence-transformers
chromadb